modified:   src/commands.py 	new file:   src/commands/__init__.py 	new file:   src/commands/base.py 	new file:   src/commands/clear_cache.py 	new file:   src/commands/help.py 	new file:   src/commands/refresh_danmaku.py
