+++ src/commands/refresh_danmaku.py
@@ -1,0 +1,482 @@
+"""
+刷新弹幕指令模块
+提供 @SXDM 指令，刷新最近播放的弹幕
+
+交互流程:
+1. @SXDM → 显示最近播放的番剧列表（#A #B #C #D #E）
+2. @SXDM #A → 显示该番剧的分集列表
+3. @SXDM #A5 → 直接触发刷新任务
+"""
+import re
+import logging
+from typing import List, Dict, TYPE_CHECKING
+from datetime import datetime
+from sqlalchemy import select, func
+from sqlalchemy.ext.asyncio import AsyncSession
+
+from .base import CommandHandler
+from .. import crud
+
+if TYPE_CHECKING:
+    from ..dandan_api import DandanSearchAnimeResponse, DandanSearchAnimeItem
+    from ..task_manager import TaskManager
+    from ..scraper_manager import ScraperManager
+    from ..rate_limiter import RateLimiter
+
+logger = logging.getLogger(__name__)
+
+
+class RefreshDanmakuCommand(CommandHandler):
+    """
+    刷新弹幕指令: @SXDM
+
+    支持三种使用方式:
+    1. @SXDM - 显示最近播放的番剧列表
+    2. @SXDM #A - 显示指定番剧的分集列表
+    3. @SXDM #A5 - 直接刷新指定番剧的第5集
+    """
+
+    # 番剧标签映射
+    ANIME_LABELS = ['#A', '#B', '#C', '#D', '#E']
+
+    def __init__(self):
+        super().__init__(
+            name="SXDM",
+            description="刷新最近播放的弹幕",
+            cooldown_seconds=2,
+            usage="@SXDM [标签] [集数]",
+            examples=[
+                "@SXDM - 查看最近播放",
+                "@SXDM #A - 查看A番剧的分集",
+                "@SXDM #A5 - 刷新A番剧第5集"
+            ]
+        )
+
+    async def execute(self, token: str, args: List[str], session: AsyncSession,
+                     config_manager, **kwargs) -> "DandanSearchAnimeResponse":
+        """执行刷新指令"""
+        from ..orm_models import Anime, AnimeSource, Episode
+        from ..task_manager import TaskManager
+        from ..scraper_manager import ScraperManager
+        from ..rate_limiter import RateLimiter
+        from .. import tasks
+
+        # 获取图片URL
+        image_url = await self.get_image_url(config_manager)
+        custom_domain = await config_manager.get("customApiDomain", "")
+
+        # 获取会话状态（用于缓存番剧和分集信息）
+        session_key = f"cmd_session_{token}"
+        session_state = await crud.get_cache(session, session_key)
+        if not session_state:
+            session_state = {}
+
+        # 阶段1: 没有参数 → 显示番剧列表
+        if not args:
+            return await self._show_anime_list(token, session, session_key, custom_domain, image_url)
+
+        # 解析参数
+        arg = args[0].upper()
+
+        # 检查参数格式
+        # 匹配 #A5 格式（标签+数字）
+        match_episode = re.match(r'^(#[A-E])(\d+)$', arg)
+        # 匹配 #A 格式（只有标签）
+        match_label = re.match(r'^#[A-E]$', arg)
+
+        if match_episode:
+            # 格式: #A5 → 直接触发刷新
+            label = match_episode.group(1)
+            episode_number = match_episode.group(2)
+
+            # 获取依赖
+            task_manager: TaskManager = kwargs.get('task_manager')
+            scraper_manager: ScraperManager = kwargs.get('scraper_manager')
+            rate_limiter: RateLimiter = kwargs.get('rate_limiter')
+
+            if not all([task_manager, scraper_manager, rate_limiter]):
+                return self.error_response(
+                    "系统依赖缺失",
+                    "无法获取必要的系统组件",
+                    image_url
+                )
+
+            return await self._trigger_refresh_by_label(
+                token, session, session_key, session_state,
+                label, episode_number,
+                task_manager, scraper_manager, rate_limiter, config_manager,
+                custom_domain, image_url
+            )
+
+        elif match_label:
+            # 格式: #A → 显示分集列表
+            return await self._show_episode_list(
+                token, session, session_key, session_state, arg, custom_domain, image_url
+            )
+
+        else:
+            # 无效格式
+            return self.error_response(
+                "无效的参数格式",
+                f"参数 '{arg}' 格式不正确\n\n💡 正确格式:\n• @SXDM #A - 查看分集列表\n• @SXDM #A5 - 刷新第5集",
+                image_url
+            )
+
+
+
+    async def _show_anime_list(
+        self,
+        token: str,
+        session: AsyncSession,
+        session_key: str,
+        custom_domain: str,
+        image_url: str
+    ) -> "DandanSearchAnimeResponse":
+        """显示最近播放的番剧列表"""
+        from ..dandan_api import DandanSearchAnimeItem
+        from ..orm_models import Anime, AnimeSource, Episode
+
+        # 读取播放历史
+        cache_key = f"play_history_{token}"
+        history = await crud.get_cache(session, cache_key)
+
+        logger.info(f"@SXDM 查询播放历史: token={token[:8]}..., cache_key={cache_key}, result={history}")
+
+        if not history:
+            history = []
+
+        if not history:
+            item = self.build_response_item(
+                anime_id=999999997,
+                title="未找到最近播放记录",
+                description="💡 提示: 播放视频后会自动记录 (10分钟有效)",
+                image_url=image_url
+            )
+            return self.build_response([item])
+
+        # 查询每部番剧的总集数
+        anime_list = []
+        for idx, record in enumerate(history[:5]):  # 只显示最近5部
+            anime_id = record["animeId"]
+
+            # 查询总集数（通过 AnimeSource 关联）
+            stmt = (
+                select(func.count(Episode.id))
+                .join(AnimeSource, Episode.sourceId == AnimeSource.id)
+                .where(AnimeSource.animeId == anime_id)
+            )
+            result = await session.execute(stmt)
+            total_episodes = result.scalar() or 0
+
+            logger.info(
+                f"@SXDM 查询番剧集数: animeId={anime_id}, "
+                f"title={record['animeTitle']}, total_episodes={total_episodes}"
+            )
+
+            anime_list.append({
+                "label": self.ANIME_LABELS[idx],
+                "animeId": anime_id,
+                "animeTitle": record["animeTitle"],
+                "totalEpisodes": total_episodes
+            })
+
+        logger.info(f"@SXDM 构建番剧列表完成: anime_list={anime_list}")
+
+        # 保存会话状态
+        session_state = {
+            "command": "SXDM",
+            "stage": "select_anime",
+            "data": {"animeList": anime_list},
+            "timestamp": datetime.now().strftime("%Y-%m-%d %H:%M:%S")
+        }
+        await crud.set_cache(session, session_key, session_state, 120)
+
+        # 记录执行时间
+        await self.record_execution(token, session)
+
+        # 构建标签列表提示
+        labels = [anime["label"] for anime in anime_list]
+        labels_text = " ".join(labels)
+
+        # 第一条：引导说明
+        anime_items = [
+            self.build_response_item(
+                anime_id=999999998,
+                title="📺 最近播放的番剧 (10分钟内)",
+                description=f"请选择要刷新的剧集作品:\n\n可用标签: {labels_text}\n\n"
+                           f"💡 使用方法:\n• @SXDM #A - 查看分集列表\n• @SXDM #A5 - 直接刷新第5集",
+                image_url=image_url,
+                episodeCount=len(anime_list)
+            )
+        ]
+
+        # 第二条开始：每部番剧
+        for anime in anime_list:
+            anime_items.append(
+                self.build_response_item(
+                    anime_id=anime["animeId"],
+                    title=f"{anime['label']} {anime['animeTitle']}",
+                    description=f"最近播放 | 共 {anime['totalEpisodes']} 集",
+                    image_url=image_url,
+                    type="tvseries",
+                    episodeCount=anime["totalEpisodes"]
+                )
+            )
+
+        logger.info(f"@SXDM 返回响应: 返回 {len(anime_items)} 条记录 (1条引导 + {len(anime_list)}部番剧)")
+
+        return self.build_response(anime_items)
+
+    async def _show_episode_list(
+        self,
+        token: str,
+        session: AsyncSession,
+        session_key: str,
+        session_state: Dict,
+        selected_label: str,
+        custom_domain: str,
+        image_url: str
+    ) -> "DandanSearchAnimeResponse":
+        """显示选中番剧的分集列表"""
+        from ..orm_models import Episode, AnimeSource
+
+        anime_list = session_state.get("data", {}).get("animeList", [])
+
+        # 查找选中的番剧
+        selected_anime = None
+        for anime in anime_list:
+            if anime["label"] == selected_label:
+                selected_anime = anime
+                break
+
+        if not selected_anime:
+            return self.error_response(
+                "无效的标签",
+                f"标签 '{selected_label}' 不存在\n💡 请输入 @SXDM 查看可用标签",
+                image_url
+            )
+
+        anime_id = selected_anime["animeId"]
+
+        # 查询分集列表（通过 AnimeSource 关联，按集数排序）
+        stmt = (
+            select(Episode)
+            .join(AnimeSource, Episode.sourceId == AnimeSource.id)
+            .where(AnimeSource.animeId == anime_id)
+            .order_by(Episode.episodeIndex)
+        )
+        result = await session.execute(stmt)
+        episodes = result.scalars().all()
+
+        if not episodes:
+            return self.error_response(
+                "未找到分集信息",
+                f"番剧 '{selected_anime['animeTitle']}' 没有分集数据",
+                image_url
+            )
+
+        # 构建分集信息（使用 Episode.commentCount 字段）
+        episode_list = []
+        for ep in episodes:
+            count = ep.commentCount or 0
+            status = "已缓存" if count > 0 else "未缓存"
+            episode_list.append({
+                "index": len(episode_list) + 1,
+                "episodeId": ep.id,
+                "episodeTitle": ep.title or f"第{ep.episodeIndex}话",
+                "commentCount": count,
+                "status": status
+            })
+
+        # 更新会话状态
+        session_state["stage"] = "select_episode"
+        session_state["data"]["selectedAnime"] = selected_anime
+        session_state["data"]["episodes"] = episode_list
+        await crud.set_cache(session, session_key, session_state, 120)
+
+        # 第一条：引导说明
+        anime_items = [
+            self.build_response_item(
+                anime_id=999999996,
+                title=f"📺 {selected_anime['animeTitle']} - 分集列表",
+                description=f"请选择要刷新的集数:\n\n共 {len(episode_list)} 集\n\n"
+                           f"💡 输入 @SXDM 标签+集数 刷新弹幕\n例如: @SXDM {selected_anime['label']}5 (刷新第5集)",
+                image_url=image_url,
+                episodeCount=len(episode_list)
+            )
+        ]
+
+        # 第二条开始：每个分集（限制显示前50集）
+        # 提取标签字母（#A -> A）
+        label_prefix = selected_anime['label'][1:]  # 去掉 # 号
+
+        for ep in episode_list[:50]:
+            anime_items.append(
+                self.build_response_item(
+                    anime_id=ep["episodeId"],
+                    title=f"[{label_prefix}{ep['index']}] {ep['episodeTitle']}",
+                    description=f"{ep['status']} | 弹幕数: {ep['commentCount']} 条",
+                    image_url=image_url,
+                    type="tvseries",
+                    episodeCount=1
+                )
+            )
+
+        logger.info(
+            f"@SXDM 返回分集列表: animeId={anime_id}, "
+            f"total={len(episode_list)}, displayed={min(50, len(episode_list))}"
+        )
+
+        return self.build_response(anime_items)
+
+    async def _trigger_refresh_by_label(
+        self,
+        token: str,
+        session: AsyncSession,
+        session_key: str,
+        session_state: Dict,
+        label: str,
+        episode_number: str,
+        task_manager,
+        scraper_manager,
+        rate_limiter,
+        config_manager,
+        custom_domain: str,
+        image_url: str
+    ) -> "DandanSearchAnimeResponse":
+        """根据标签和集数触发刷新任务（格式: #A5）"""
+        from ..orm_models import Episode, AnimeSource
+        from .. import tasks
+
+        # 解析集数编号
+        try:
+            ep_num = int(episode_number)
+        except ValueError:
+            return self.error_response(
+                "无效的集数",
+                f"集数 '{episode_number}' 不是有效的数字",
+                image_url
+            )
+
+        # 从播放历史中获取番剧列表
+        cache_key = f"play_history_{token}"
+        history = await crud.get_cache(session, cache_key)
+        if not history:
+            return self.error_response(
+                "未找到播放历史",
+                "请先播放视频后再使用刷新功能",
+                image_url
+            )
+
+        # 构建标签到番剧的映射
+        anime_list = []
+        for idx, record in enumerate(history[:5]):
+            anime_list.append({
+                "label": self.ANIME_LABELS[idx],
+                "animeId": record["animeId"],
+                "animeTitle": record["animeTitle"]
+            })
+
+        # 查找对应标签的番剧
+        selected_anime = None
+        for anime in anime_list:
+            if anime["label"] == label:
+                selected_anime = anime
+                break
+
+        if not selected_anime:
+            labels = " ".join([a["label"] for a in anime_list])
+            return self.error_response(
+                "无效的标签",
+                f"标签 '{label}' 不存在\n💡 可用标签: {labels}",
+                image_url
+            )
+
+        anime_id = selected_anime["animeId"]
+        anime_title = selected_anime["animeTitle"]
+
+        # 查询该番剧的所有分集（按集数排序）
+        stmt = (
+            select(Episode)
+            .join(AnimeSource, Episode.sourceId == AnimeSource.id)
+            .where(AnimeSource.animeId == anime_id)
+            .order_by(Episode.episodeIndex)
+        )
+        result = await session.execute(stmt)
+        episodes = result.scalars().all()
+
+        if not episodes:
+            return self.error_response(
+                "未找到番剧分集信息",
+                f"番剧: {anime_title}",
+                image_url
+            )
+
+        # 验证集数编号
+        if ep_num < 1 or ep_num > len(episodes):
+            return self.error_response(
+                "无效的集数",
+                f"集数: {ep_num}\n番剧: {anime_title}\n可用集数: 1-{len(episodes)}",
+                image_url
+            )
+
+        # 获取对应分集
+        selected_episode = episodes[ep_num - 1]
+        episode_id = selected_episode.id
+        episode_title = selected_episode.title or f"第{selected_episode.episodeIndex}话"
+
+        # 验证分集存在
+        info = await crud.get_episode_for_refresh(session, episode_id)
+        if not info:
+            return self.error_response(
+                "分集信息异常",
+                f"分集: {episode_title}",
+                image_url
+            )
+
+        # 提交刷新任务
+        try:
+            unique_key = f"refresh-episode-{episode_id}"
+
+            task_id, _ = await task_manager.submit_task(
+                lambda s, cb: tasks.refresh_episode_task(
+                    episode_id, s, scraper_manager, rate_limiter, cb, config_manager
+                ),
+                f"指令刷新: {anime_title} - {episode_title}",
+                unique_key=unique_key
+            )
+
+            # 记录执行时间
+            await self.record_execution(token, session)
+
+            logger.info(
+                f"@SXDM 提交刷新任务: label={label}, episode_number={ep_num}, "
+                f"episodeId={episode_id}, anime={anime_title}, taskId={task_id}"
+            )
+
+            message = (
+                f"✓ 刷新任务已提交\n\n"
+                f"番剧: {anime_title}\n"
+                f"分集: [{ep_num}] {episode_title}\n"
+                f"任务ID: {task_id}\n\n"
+                f"🔄 任务处理中，请稍候15秒后重新获取弹幕"
+            )
+
+            item = self.build_response_item(
+                anime_id=999999995,
+                title="✓ 弹幕刷新任务已提交",
+                description=message,
+                image_url=image_url
+            )
+
+            return self.build_response([item])
+
+        except Exception as e:
+            logger.error(f"@SXDM 提交刷新任务失败: {e}", exc_info=True)
+            return self.error_response(
+                "任务提交失败",
+                f"错误详情: {str(e)}",
+                image_url
+            )
+
+
